Minor bugs fixes still remaining

+++ backend/src/main/java/com/example/gov_scheme_backend/dto/request/workflow/WorkflowActionRequest.java
@@ -4,6 +4,8 @@ import com.example.gov_scheme_backend.enums.WorkflowAction;
 import jakarta.validation.constraints.NotNull;
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 public class WorkflowActionRequest {
 
@@ -13,7 +15,7 @@ public class WorkflowActionRequest {
     private String remarks;
 
     // Used only when Finance Officer approves
-    private Double approvedAmount;
+    private BigDecimal approvedAmount;
 
     // Used only when Finance Officer approves — how many installments to split the grant into
     private Integer numberOfInstallments;
